backend/requirements.txt
# Core Django
# Use Django 4.2 LTS series for compatibility with currently pinned dependencies
# (django-celery-beat and others may not yet support Django 5.x in this requirements set)
Django>=4.2,<5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
argon2-cffi==23.1.0

# Async Tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage
boto3==1.34.34
django-storages==1.14.2
Pillow==10.2.0

# Payments
stripe==8.2.0

# AI/ML
openai==2.8.0
spacy==3.7.4
sentence-transformers==2.5.1
PyPDF2==3.0.1
python-docx==1.1.0

# API Documentation
drf-spectacular==0.27.1

# Monitoring & Logging
sentry-sdk==1.40.3

# Environment
python-decouple==3.8
python-dotenv==1.0.1

# Production Server
gunicorn==21.2.0

# Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.6.0

# Code Quality
black==24.2.0
flake8==7.0.0
isort==5.13.2

# Utilities
pytz==2024.1
python-dateutil==2.8.2
requests==2.31.0
